requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Base de datos
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Validación y serialización
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Caché y optimización
redis==5.0.1
prometheus-client==0.19.0
psutil==5.9.7

# Seguridad y autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilidades
python-magic==0.4.27
bleach==6.1.0
python-dotenv==1.0.0 